move functions to device
