+++ authentik/stages/authenticator_mobile/models.py
@@ -87,3 +87,5 @@ class MobileDeviceToken(ExpiringModel):
     device = models.ForeignKey(MobileDevice, on_delete=models.CASCADE, null=True)
     user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
     token = models.TextField(default=default_token_key)
+
+    firebase_token = models.TextField(blank=True)
